Add files via upload
In this mini project I had dataset of patients who responded to different medication like DRUG A, DRUG B, DRUG C and DRUG D, which acted like target class for pour model. I had BP, AGE  and Cholesterol acting as features which where used to train the model on 80% dataset and 20% where used for testing. 

I obtained an accuracy of 100% in this project.

Regards
Gaurav R